+++ CHANGELOG.md
@@ -7,6 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2026-09-02
+
+### Changed
+
+- let Dependabot PRs patch-bump
+
+## [1.0.0] - 2026-09-02
+
+### Added
+
+- lockstep npm packages to 1.0.0
+
+### Changed
+
+- Major reset to 1.0.0 via majorEnv
+
+### Fixed
+
+- update BLOCKERS.md test
+
 ## [0.1.0] - 2026-08-22
 
 ### Added